Added listFailedActions method
The listFailedActions acepts a string[] as input, which can be emnoty, and check if any of the cids in the input array have failed being pushed to the backend. It lists all failed cids if input is an empty array.
If any of hte iput cids are failed, then the putput shows tehm. If none failed, output is false.